backend/pyproject.toml
[tool.poetry]
name = "therix-chatbot"
version = "0.1.0"
description = ""
authors = ["Coditas LLP"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
therix = "^0.1.26"
elevenlabs = "^1.5.0"
mangum = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
